Restrictions not retrieved in every activity creation now
